Moved tests into its own package
